# docs/swagger.yaml
basePath: /todos
definitions:
  models.Todo:
    properties:
      completed:
        type: boolean
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  models.TodoCompletedStatus:
    properties:
      completed:
        type: boolean
    type: object
info:
  contact:
    email: amadijerry823@gmail.com
    name: Amadi Jerry
    url: https://twitter.com/Jerry_Jaeger_
  description: A TODO APP API in Go using Gin framework.
  title: Todo App API
  version: "1.0"
paths:
...
  /status/{id}:
    patch:
      description: Change the status of a Todo from uncompleted to completed and vice
        versa
      parameters:
      - description: update Todo Status by Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: change the status of a Todo
        in: body
        name: completed
        required: true
        schema:
          $ref: '#/definitions/models.TodoCompletedStatus'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Update Todo Status
      tags:
      - Todos
...
swagger: "2.0"